Deleted useless methods and rewrited to work only from index to simple add/remode operations

+++ BestStudentCafedra/Controllers/ProposedTopicController.cs
@@ -20,38 +20,12 @@ namespace BestStudentCafedra.Controllers
         }
 
         // GET: ProposedTopic
-        public async Task<IActionResult> Index()
+        public async Task<IActionResult> Index(ProposedTopic topic)
         {
-            return View(await _context.ProposedTopics.ToListAsync());
+            ViewData["topics"] = await _context.ProposedTopics.ToListAsync();
+            return View(topic == null? new ProposedTopic(): topic);
         }
 
-        // GET: ProposedTopic/Details/5
-        public async Task<IActionResult> Details(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var proposedTopic = await _context.ProposedTopics
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (proposedTopic == null)
-            {
-                return NotFound();
-            }
-
-            return View(proposedTopic);
-        }
-
-        // GET: ProposedTopic/Create
-        public IActionResult Create()
-        {
-            return View();
-        }
-
-        // POST: ProposedTopic/Create
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create([Bind("Id,Name")] ProposedTopic proposedTopic)
@@ -62,76 +36,9 @@ namespace BestStudentCafedra.Controllers
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
-            return View(proposedTopic);
-        }
-
-        // GET: ProposedTopic/Edit/5
-        public async Task<IActionResult> Edit(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var proposedTopic = await _context.ProposedTopics.FindAsync(id);
-            if (proposedTopic == null)
-            {
-                return NotFound();
-            }
-            return View(proposedTopic);
-        }
-
-        // POST: ProposedTopic/Edit/5
-        // To protect from overposting attacks, enable the specific properties you want to bind to.
-        // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id,Name")] ProposedTopic proposedTopic)
-        {
-            if (id != proposedTopic.Id)
-            {
-                return NotFound();
-            }
-
-            if (ModelState.IsValid)
-            {
-                try
-                {
-                    _context.Update(proposedTopic);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!ProposedTopicExists(proposedTopic.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
-                return RedirectToAction(nameof(Index));
-            }
-            return View(proposedTopic);
-        }
-
-        // GET: ProposedTopic/Delete/5
-        public async Task<IActionResult> Delete(int? id)
-        {
-            if (id == null)
-            {
-                return NotFound();
-            }
-
-            var proposedTopic = await _context.ProposedTopics
-                .FirstOrDefaultAsync(m => m.Id == id);
-            if (proposedTopic == null)
-            {
-                return NotFound();
-            }
-
-            return View(proposedTopic);
+            ModelState.AddModelError("", "Длина темы должна быть больше 10 символов");
+            ViewData["topics"] = await _context.ProposedTopics.ToListAsync();
+            return View("Index", proposedTopic);
         }
 
         // POST: ProposedTopic/Delete/5
@@ -139,15 +46,15 @@ namespace BestStudentCafedra.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var proposedTopic = await _context.ProposedTopics.FindAsync(id);
+            var proposedTopic = await _context.ProposedTopics
+                .FirstOrDefaultAsync(m => m.Id == id);
+            if (proposedTopic == null)
+            {
+                return NotFound();
+            }
             _context.ProposedTopics.Remove(proposedTopic);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
-        }
-
-        private bool ProposedTopicExists(int id)
-        {
-            return _context.ProposedTopics.Any(e => e.Id == id);
         }
     }
 }
